fix: rtk implementation
